+++ src/common/Panel.tsx
